web/requirements.txt
flask
psycopg2
uwsgi
python-dotenv